For some reason the compiler changed currently using Java 1.8.0_121. Although the compiler *May be using 1.7 Regardless the changes this code currently runs using 1.7+

+++ src/AdjacencyMatrix.java
@@ -89,7 +89,7 @@ public class AdjacencyMatrix {
                     if (w==null){
                         row[j]=INF;
                     } else {
-                        row[j]=(int)w;
+                        row[j]=((Integer)w).intValue();
                     }
                 }
             }
